terry.framework: Fix wrong class name for Controller and Service
